Add files via upload
This is the FoodItem class with a constructor to initialize a food item. Also it accepts as input a food item name, fat, carbs, and protein and the number of servings. The program creates a food item using the constructor parameters' default values and a food item using the input values